Add files via upload
Stability Study in Power Electrical Systems using Nonlinear Models

Project Description:
This project focuses on the analysis of the stability of power electrical systems, specifically synchronous machines, through the utilization of nonlinear models. The research aims to understand the dynamic behavior of these machines under various disturbance conditions and variations in mechanical power.
The main objective is to evaluate how stability parameters, such as the time constant of the axis \( H \) and the damping factor, affect the system's response. A comparative analysis between linear and nonlinear models will be conducted to identify the differences in stability and performance under varying operating conditions.

Methodology:
The methodology includes simulating synchronous machines using programming tools in Python, specifically the `scipy.integrate.odeint` library to solve differential equations. The results will be visualized through interactive graphs generated with Plotly, allowing for clear and effective interpretation of the data.

Expected Outcomes:
This study aims to contribute to the development of more sophisticated control strategies in power electrical systems, promoting greater efficiency and stability in the operation of these critical systems.